Separate reentrancy tests
