+++ radix-engine/tests/blueprints/component/src/lib.rs
@@ -5,4 +5,3 @@ pub mod component;
 pub mod cross_component;
 pub mod external_blueprint_target;
 pub mod mutable_access_rules;
-pub mod reentrant_component;
