+++ radix-engine/tests/blueprints/reentrancy/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "reentrancy"
+version = "0.7.0"
+edition = "2021"
+
+[dependencies]
+sbor = { path = "../../../../sbor" }
+scrypto = { path = "../../../../scrypto" }
+
+[dev-dependencies]
+radix-engine = { path = "../../../../radix-engine" }
+
+[lib]
+crate-type = ["cdylib", "lib"]
