+++ radix-engine/tests/component.rs
@@ -1,6 +1,6 @@
 use radix_engine::engine::{
-    ApplicationError, AuthError, InterpreterError, KernelError, LockState, ModuleError,
-    RuntimeError, ScryptoFnResolvingError, TrackError,
+    ApplicationError, AuthError, InterpreterError, ModuleError,
+    RuntimeError, ScryptoFnResolvingError,
 };
 use radix_engine::ledger::TypedInMemorySubstateStore;
 use radix_engine::model::AccessRulesChainError;
@@ -95,112 +95,6 @@ fn invalid_blueprint_name_should_cause_error() {
         } else {
             false
         }
-    });
-}
-
-#[test]
-fn mut_reentrancy_should_not_be_possible() {
-    // Arrange
-    let mut store = TypedInMemorySubstateStore::with_bootstrap();
-    let mut test_runner = TestRunner::new(true, &mut store);
-    let package_address = test_runner.compile_and_publish("./tests/blueprints/component");
-    let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
-        .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_function(package_address, "ReentrantComponent", "new", args!())
-        .build();
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-    receipt.expect_commit_success();
-    let component_address = receipt
-        .expect_commit()
-        .entity_changes
-        .new_component_addresses[0];
-
-    // Act
-    let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
-        .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_method(component_address, "call_mut_self", args!())
-        .build();
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-
-    // Assert
-    receipt.expect_specific_failure(|e| {
-        matches!(
-            e,
-            RuntimeError::KernelError(KernelError::TrackError(TrackError::SubstateLocked(
-                SubstateId(
-                    RENodeId::Component(..),
-                    SubstateOffset::Component(ComponentOffset::State)
-                ),
-                LockState::Write
-            )))
-        )
-    });
-}
-
-#[test]
-fn read_reentrancy_should_be_possible() {
-    // Arrange
-    let mut store = TypedInMemorySubstateStore::with_bootstrap();
-    let mut test_runner = TestRunner::new(true, &mut store);
-    let package_address = test_runner.compile_and_publish("./tests/blueprints/component");
-    let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
-        .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_function(package_address, "ReentrantComponent", "new", args!())
-        .build();
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-    receipt.expect_commit_success();
-    let component_address = receipt
-        .expect_commit()
-        .entity_changes
-        .new_component_addresses[0];
-
-    // Act
-    let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
-        .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_method(component_address, "call_self", args!())
-        .build();
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-
-    // Assert
-    receipt.expect_commit_success();
-}
-
-#[test]
-fn read_then_mut_reentrancy_should_not_be_possible() {
-    // Arrange
-    let mut store = TypedInMemorySubstateStore::with_bootstrap();
-    let mut test_runner = TestRunner::new(true, &mut store);
-    let package_address = test_runner.compile_and_publish("./tests/blueprints/component");
-    let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
-        .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_function(package_address, "ReentrantComponent", "new", args!())
-        .build();
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-    receipt.expect_commit_success();
-    let component_address = receipt
-        .expect_commit()
-        .entity_changes
-        .new_component_addresses[0];
-
-    // Act
-    let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
-        .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_method(component_address, "call_mut_self_2", args!())
-        .build();
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-
-    // Assert
-    receipt.expect_specific_failure(|e| {
-        matches!(
-            e,
-            RuntimeError::KernelError(KernelError::TrackError(TrackError::SubstateLocked(
-                SubstateId(
-                    RENodeId::Component(..),
-                    SubstateOffset::Component(ComponentOffset::State)
-                ),
-                LockState::Read(1),
-            )))
-        )
     });
 }
